MedicalLap/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
configparser==5.3.0
contourpy==1.0.6
cycler==0.11.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
configparser==5.3.0
contourpy==1.0.6
cycler==0.11.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
shutils==0.1.0
six==1.16.0
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0